+++ src/ts/components/Container.ts
@@ -40,7 +40,7 @@ export class Container<Config extends ContainerConfig> extends Component<Config>
   /**
    * A reference to the inner element that contains the components of the container.
    */
-  private innerContainerElement: DOM;
+  protected innerContainerElement: DOM;
   private componentsToAppend: Component<ComponentConfig>[];
   private componentsToPrepend: Component<ComponentConfig>[];
   private componentsToRemove: Component<ComponentConfig>[];